# lib/api-zod/src/generated/types/cart.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface Cart {
  id: string;
  items: CartItem[];
  /**
     * Total units across all lines
     * @minimum 0
     */
  item_count: number;
  /** @minimum 0 */
  subtotal_da: number;
}