# lib/api-zod/src/generated/types/placeOrderRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlaceOrderRequest {
  cart_id: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  contact_name: string;
  /** @pattern ^\+?[0-9 ()-]{6,20}$ */
  contact_phone: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  shipping_address: string;
  /** @maxLength 500 */
  customer_note?: string;
}