# lib/api-zod/src/generated/types/updateCartItemRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCartItemRequest {
  /**
     * 0 removes the line
     * @minimum 0
     * @maximum 99
     */
  quantity: number;
}